4.1.0 (#520)
* Sample solution.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update modules/config.py

Increase efficiency by using a boolean in place of a functional check.


* Bump softprops/action-gh-release from 1 to 2

Bumps [softprops/action-gh-release](https://github.com/softprops/action-gh-release) from 1 to 2.
- [Release notes](https://github.com/softprops/action-gh-release/releases)
- [Changelog](https://github.com/softprops/action-gh-release/blob/master/CHANGELOG.md)
- [Commits](https://github.com/softprops/action-gh-release/compare/v1...v2)

---
updated-dependencies:
- dependency-name: softprops/action-gh-release
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* update group_uplolad_speed logic

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Bump dependabot/fetch-metadata from 1.6.0 to 2.0.0

Bumps [dependabot/fetch-metadata](https://github.com/dependabot/fetch-metadata) from 1.6.0 to 2.0.0.
- [Release notes](https://github.com/dependabot/fetch-metadata/releases)
- [Commits](https://github.com/dependabot/fetch-metadata/compare/v1.6.0...v2.0.0)

---
updated-dependencies:
- dependency-name: dependabot/fetch-metadata
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Bump pre-commit from 3.6.2 to 3.7.0

Bumps [pre-commit](https://github.com/pre-commit/pre-commit) from 3.6.2 to 3.7.0.
- [Release notes](https://github.com/pre-commit/pre-commit/releases)
- [Changelog](https://github.com/pre-commit/pre-commit/blob/main/CHANGELOG.md)
- [Commits](https://github.com/pre-commit/pre-commit/compare/v3.6.2...v3.7.0)

---
updated-dependencies:
- dependency-name: pre-commit
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* [pre-commit.ci] pre-commit autoupdate

updates:
- [github.com/hhatto/autopep8: v2.0.4 → v2.1.0](https://github.com/hhatto/autopep8/compare/v2.0.4...v2.1.0)
- [github.com/asottile/pyupgrade: v3.15.1 → v3.15.2](https://github.com/asottile/pyupgrade/compare/v3.15.1...v3.15.2)
- [github.com/psf/black: 24.2.0 → 24.3.0](https://github.com/psf/black/compare/24.2.0...24.3.0)

* Bump qbittorrent-api from 2024.2.59 to 2024.3.60

Bumps [qbittorrent-api](https://github.com/rmartin16/qbittorrent-api) from 2024.2.59 to 2024.3.60.
- [Release notes](https://github.com/rmartin16/qbittorrent-api/releases)
- [Changelog](https://github.com/rmartin16/qbittorrent-api/blob/main/CHANGELOG.md)
- [Commits](https://github.com/rmartin16/qbittorrent-api/compare/v2024.2.59...v2024.3.60)

---
updated-dependencies:
- dependency-name: qbittorrent-api
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump 4.0.10

* Revert "Add Aggregate Group Speed Capabilities"

* Bump gitpython from 3.1.42 to 3.1.43

Bumps [gitpython](https://github.com/gitpython-developers/GitPython) from 3.1.42 to 3.1.43.
- [Release notes](https://github.com/gitpython-developers/GitPython/releases)
- [Changelog](https://github.com/gitpython-developers/GitPython/blob/main/CHANGES)
- [Commits](https://github.com/gitpython-developers/GitPython/compare/3.1.42...3.1.43)

---
updated-dependencies:
- dependency-name: gitpython
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Add new args for mover.old

* Revert "Revert "Add Aggregate Group Speed Capabilities""

This reverts commit f112b2c92e41f3a80090fb3ac5b889902b042367.
Fixes bug in original commit code

* Remove username and password from debug logs sinced they are redacted anyways

* adds support for #513

* Fixes #501

* Fixes bug not logging torrents being deleted after meeting share_limits

* Adds additional error checking for invalid share_limits in config

* fixes bug in torrents not untagging min_seeding_time_tag

* Fixes bug in #494

* Bi-directional Wiki Sync Action

* Bump actions/checkout from 2 to 4

Bumps [actions/checkout](https://github.com/actions/checkout) from 2 to 4.
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/v2...v4)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Bug fix on share limits being applied every run

* Revert "Revert "Revert "Add Aggregate Group Speed Capabilities"""

This reverts commit 41d5170ba649f47fe4ff8fd59a087453bd524089.

* Fixes #494 with new enable_group_upload_speed flag

* Fixes bug found in #494

* 4.1.0

---------

+++ modules/util.py
@@ -421,7 +421,7 @@ def copy_files(src, dest):
         logger.error(ex)
 
 
-def remove_empty_directories(pathlib_root_dir, pattern):
+def remove_empty_directories(pathlib_root_dir, pattern, excluded_paths=None):
     """Remove empty directories recursively."""
     pathlib_root_dir = Path(pathlib_root_dir)
     try:
@@ -435,6 +435,8 @@ def remove_empty_directories(pathlib_root_dir, pattern):
         longest.append(pathlib_root_dir)  # delete the folder itself if it's empty
         for pdir in longest:
             try:
+                if str(pdir) in excluded_paths:
+                    continue
                 pdir.rmdir()  # remove directory if empty
             except (FileNotFoundError, OSError):
                 continue  # catch and continue if non-empty, folders within could already be deleted if run in parallel
